cluster/applications/slack-bolt-proxy/pyproject.toml
[tool.poetry]
name = "slack-bolt-proxy"
version = "0.1.0"
description = ""
authors = ["Kai Aihara <kaidotio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"

asyncio = "^3.4.3"
aiohttp = "^3.8.4"

slack-bolt = "^1.17.0"

pydantic = "^1.10.10"

opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-exporter-otlp = "^1.24.0"
opentelemetry-instrumentation-aiohttp-client = "^0.45b0"
opentelemetry-exporter-prometheus = "^0.45b0"
prometheus-client = "^0.17.0"
python-dotenv = "^1.0.0"
python-json-logger = "^2.0.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
